requirements.txt
numpy>=1.22.0
matplotlib>=3.5.1
Pillow>=9.0.0
tqdm>=4.62.3